Refactor GameFormComponent for improved tag handling and error management
- Updated comments for clarity on tag filtering process.
- Simplified tag input handling by removing unnecessary clicks and focusing on predefined tags.
- Enhanced error handling by ensuring input is cleared on failure.
- Reduced wait times for UI updates to optimize performance.
